+++ src/LocalFilesystem.php
@@ -24,12 +24,35 @@ class LocalFilesystem extends Filesystem
     public $path;
 
     /**
+     * @var int
+     */
+    public $writeFlags = LOCK_EX;
+
+    /**
+     * @var int
+     */
+    public $linkHandling = Local::DISALLOW_LINKS;
+
+    /**
+     * @var array
+     */
+    public $permissions = [];
+
+    /**
      * @inheritdoc
      */
     public function init()
     {
         if ($this->path === null) {
             throw new InvalidConfigException('The "path" property must be set.');
+        }
+
+        if (!in_array($this->writeFlags, [0, LOCK_SH, LOCK_EX, LOCK_UN, LOCK_NB], true)) {
+            throw new InvalidConfigException('The "writeFlags" property value is invalid.');
+        }
+
+        if ($this->linkHandling !== Local::DISALLOW_LINKS && $this->linkHandling !== Local::SKIP_LINKS) {
+            throw new InvalidConfigException('The "linkHandling" property value is invalid.');
         }
 
         $this->path = Yii::getAlias($this->path);
@@ -42,6 +65,7 @@ class LocalFilesystem extends Filesystem
      */
     protected function prepareAdapter()
     {
-        return new Local($this->path);
+        return new Local($this->path, $this->writeFlags, $this->linkHandling, $this->permissions);
     }
 }
+